requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
diskcache==5.6.3
fastapi==0.115.5
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
llama_cpp_python==0.3.2
MarkupSafe==3.0.2
numpy==2.1.3
packaging==24.2
pluggy==1.5.0
pydantic==2.10.1
pydantic_core==2.27.1
pytest==8.3.3
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
